Move SentryMixin to a separate file.
